Some performance optimisations. Row counts are tracked rather than calculated each time.
Adding and removing rows now fires a rowCountChanged event and an
internal rowHeight changed event which avoids the need to recalculate
the viewport - just the v scrollbar maximum.

The content provider returns children as a list rather than an array.

It's now possible to detect the row number cells are being tracked in
the mouse listener.